refactor: remove shared/lib/selectors/util.ts re-export
