--- shared/lib/selectors/util.ts
@@ -1,5 +1,0 @@
-// re-export for backward compatibility
-export {
-  createDeepEqualSelector,
-  createParameterizedSelector,
-} from './selector-creators';
