+++ ui/selectors/transactions.js
@@ -21,7 +21,7 @@ import {
   getProviderConfig,
   getCurrentChainId,
 } from '../../shared/lib/selectors/networks';
-import { createDeepEqualSelector } from '../../shared/lib/selectors/util';
+import { createDeepEqualSelector } from '../../shared/lib/selectors/selector-creators';
 import {
   createShallowEqualInputAndResultSelector,
   createParameterizedShallowEqualSelector,
